\name{EcoinfPubsRpackage-package}
\alias{EcoinfPubsRpackage-package}
\alias{EcoinfPubsRpackage}
\docType{package}
\title{ EcoinfPubsRpackage }
\description{ Produce reports of Ecoinformatics and Biodiversity publications. }
\details{
Use this package to produce reports of Ecoinformatics and Biodiversity publications that are entered into the publication database though the GoogleForm here: \url{https://goo.gl/forms/waQdYPGjeIDWN4r03}.

The package also facilitates automatically emailing people to (1) add DOI's to their publications, and (2) update the status of a publications from 'Submitted' to 'Accepted'.
}
